+++ src/main/java/com/atlassian/performance/tools/awsinfrastructure/api/dataset/DatasetHost.java
@@ -1,0 +1,12 @@
+package com.atlassian.performance.tools.awsinfrastructure.api.dataset;
+
+import com.atlassian.performance.tools.awsinfrastructure.api.InfrastructureFormula;
+import com.atlassian.performance.tools.infrastructure.api.dataset.Dataset;
+
+/**
+ * @since 2.12.0
+ */
+public interface DatasetHost {
+
+    InfrastructureFormula<?> host(Dataset dataset);
+}