+++ CHANGELOG.md
@@ -26,6 +26,7 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ### Added
 - Makes ApacheEc2LoadBalancerFormula available as part of API
 - Let `AwsDatasetModification` build the `Dataset` without disturbing the `InfrastructureFormula` and vice versa.
+- Add `SshMysqlDatasetPublication`. Resolve [JPERF-519].
 
 ### Fixed
 - Error while AwsDatasetModification saves dataset. Fix [JPERF-515].
@@ -36,6 +37,7 @@ Dropping a requirement of a major version of a dependency is a new contract.
 ### Deprecated
 - Replace `AwsDatasetModification.formula` with a pair of `dataset` and `host`.
 
+[JPERF-519]: https://ecosystem.atlassian.net/browse/JPERF-519
 [JPERF-515]: https://ecosystem.atlassian.net/browse/JPERF-515
 [JPERF-506]: https://ecosystem.atlassian.net/browse/JPERF-506
 [JPERF-406]: https://ecosystem.atlassian.net/browse/JPERF-406