+++ CHANGELOG.md
@@ -21,7 +21,10 @@ Adding a requirement of a major version of a dependency is breaking a contract.
 Dropping a requirement of a major version of a dependency is a new contract.
 
 ## [Unreleased]
-[Unreleased]: https://github.com/atlassian/aws-infrastructure/compare/release-2.11.0...master
+[Unreleased]: https://github.com/atlassian/aws-infrastructure/compare/release-2.12.0...master
+
+## [2.12.0] - 2019-07-08
+[2.12.0]: https://github.com/atlassian/aws-infrastructure/compare/release-2.11.0...release-2.12.0
 
 ### Added
 - Makes ApacheEc2LoadBalancerFormula available as part of API